pyproject.toml
[project]
name = "data-product-strands-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "botocore>=1.40.34",
    "pandas>=2.3.2",
    "strands-agents>=1.9.0",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.2.8",
]
